Lots of QOI improvements:
 - Add a status command
 - Allow the runner to gracefully terminate if it's in the middle of a
   job, rather than killing it immediately
 - Fix some issues with the atomic update.  The runner will now always
   finish currently executing jobs and then pick up the new base image
   on the next run.

+++ run-actions-runner.sh
@@ -1,11 +1,25 @@
 #!/bin/sh
 set -euo pipefail
-RUNNER_CLONE_NAME=${RUNNER_NAME}-$(uuidgen)
+if [ -f /var/run/github-runners.${RUNNER_NAME} ]; then
+	echo ${RUNNER_NAME} already running.
+	echo Please delete /var/run/github-runners.${RUNNER_NAME} if the system did not gracefully shut down.
+	exit 1
+fi
+RUNNER_CLONE_NAME=${RUNNER_NAME}-ephemeral
+cleanup() {
+	echo Signal received, exiting...
+	pot stop -p ${RUNNER_CLONE_NAME}
+	pot destroy -p ${RUNNER_CLONE_NAME}
+	rm /var/run/github-runners.${RUNNER_NAME}
+}
+trap cleanup INT
 echo $$ > /var/run/github-runners.${RUNNER_NAME}
 while [ -f /var/run/github-runners ] ; do
+	if [ "$(cat /var/run/github-runners.${RUNNER_NAME})" != $$ ] ; then
+		exit 0
+	fi
 	# Acquire a lock while cloning to prevent races against an update of the base image. 
-	lockf -k /var/run/github-runners.${RUNNER_NAME}.lock pot clone -P ${RUNNER_NAME} -p ${RUNNER_CLONE_NAME}
+	lockf -k /var/run/github-runners.${RUNNER_NAME}.lock pot clone -F -P ${RUNNER_NAME} -p ${RUNNER_CLONE_NAME}
 	pot start -p ${RUNNER_CLONE_NAME}
-	pot stop -p ${RUNNER_CLONE_NAME}
 	pot destroy -p ${RUNNER_CLONE_NAME}
 done
